rutube_site/requirements.txt
scikit-learn==1.5.2
sentence-transformers==3.0.1
channels==4.1.0
daphne==4.1.2
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
drfpasswordless==1.5.9
pillow==10.3.0
PyJWT==1.7.1
gradio-client==1.3.0
numpy==1.26.4
aiogram==3.13.0
pydub==0.25.1
requests==2.32.3
beautifulsoup4==4.12.3
python-docx==1.1.2
python-environ==0.4.54
psycopg2-binary==2.9.9
pyTelegramBotAPI==4.22.1
celery==5.4.0
redis==3.4.1
flower==2.0.1